Move account header and row to dedicate files
