+++ src/components/screens/monitor/accounts/tableHeader.js
@@ -1,0 +1,24 @@
+import grid from 'flexboxgrid/dist/flexboxgrid.css';
+
+export default [
+  {
+    title: 'Rank',
+    classList: `${grid['col-xs-1']} ${grid['col-md-1']}`,
+  },
+  {
+    title: 'Address',
+    classList: `${grid['col-xs-3']} ${grid['col-md-5']}`,
+  },
+  {
+    title: 'Balance',
+    classList: `${grid['col-xs-3']} ${grid['col-md-3']}`,
+  },
+  {
+    title: 'Supply',
+    classList: `${grid['col-xs-2']} ${grid['col-md-1']}`,
+  },
+  {
+    title: 'Owner',
+    classList: `${grid['col-xs-3']} ${grid['col-md-2']}`,
+  },
+];
